% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{copy_shmi_example}
\alias{copy_shmi_example}
\title{Copy the example SHMI Excel file to a local path}
\usage{
copy_shmi_example(path = "SHMI_example_1.xlsx")
}
\arguments{
\item{path}{File path to save the example.}
}
\value{
The path to the saved file.
}
\description{
Copy the example SHMI Excel file to a local path
}
