% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_shmi_template}
\alias{get_shmi_template}
\title{Get the SHMI Excel template}
\usage{
get_shmi_template(path = "SHMI_template.xlsx")
}
\arguments{
\item{path}{File path where the template should be saved.}
}
\value{
The path to the saved file.
}
\description{
Copies the SHMI Excel template to a user-specified location.
}
