% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shmi_lollipop.R
\name{plot_shmi_lollipop}
\alias{plot_shmi_lollipop}
\title{Plot SHMI across management units}
\usage{
plot_shmi_lollipop(shmi)
}
\arguments{
\item{shmi}{A data frame with \code{MGT_combo} and \code{SHMI}, such as
\code{build_shmi()$indicator_df}.}
}
\value{
A ggplot object.
}
\description{
Horizontal lollipop chart of overall SHMI, one line per management unit,
sorted by score.
}
\examples{
scores <- data.frame(MGT_combo = c("field_1", "field_2", "field_3"),
                     SHMI = c(62.3, 48.1, 75.6))
plot_shmi_lollipop(scores)

}
\seealso{
\code{\link[=plot_shmi_gauge]{plot_shmi_gauge()}}, \code{\link[=build_shmi]{build_shmi()}}
}
