% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_shmi_example}
\alias{get_shmi_example}
\title{Load the example SHMI Excel file}
\usage{
get_shmi_example()
}
\value{
The path to the example Excel file.
}
\description{
Provides a fully populated example dataset for testing the SHMI workflow.
}
